Add Logic support.
A note for logic users:

There is a small bit of extra code required in order to pass the AU validation steps in Logic.

Logic will attempt to scan your audio plugins, and check them at various channel set options. Because we're coding the plugin in stereo, you'll hit issues with mono configurations in Logic. This isn't an issue in Ableton, because Ableton runs stereo only. In order to fix this, you'll need to specify your channel configuration.

This is done in two places. In the construction of your audio plugin, you'll specify the number of input and output buses. We can clean this code up to look like this:

KadenzeAudioProcessor::KadenzeAudioProcessor()
     : AudioProcessor (BusesProperties()
                       .withInput  ("Input",  AudioChannelSet::stereo(), true)
                       .withOutput ("Output", AudioChannelSet::stereo(), true))
{
This is stating we have 2 input, and 2 outputs.
When Logic checks the information about your application, it will call the isBusesLayoutSupported function. You should edit this to only accept stereo configurations. Something like this will work:

bool KadenzeAudioProcessor::isBusesLayoutSupported (const BusesLayout& layouts) const
{
    if (layouts.getMainInputChannelSet() == AudioChannelSet::stereo() &&
        layouts.getMainOutputChannelSet() == AudioChannelSet::stereo()) {
        return true;
    } else {
        return false;
    }
}
Happy coding!